+++ name.abuchen.portfolio/src/name/abuchen/portfolio/online/ImportFinanzenNetQuotes.java
@@ -106,13 +106,13 @@ public class ImportFinanzenNetQuotes
         }
     }
 
-    private int parsePrice(String text) throws IOException
+    private long parsePrice(String text) throws IOException
     {
         try
         {
             DecimalFormat fmt = new DecimalFormat("#,##0.###", new DecimalFormatSymbols(Locale.GERMANY)); //$NON-NLS-1$
             Number q = fmt.parse(text);
-            return (int) (q.doubleValue() * 100);
+            return (long) (q.doubleValue() * 100);
         }
         catch (ParseException e)
         {
